phase(1): the hashmap-surface flip — the public keyed-collection surface becomes ONE family (HashMap<K, V> / HashSet<T>, the val type selecting the storage) through generalized RFC 0043 alias ROWS, engine and every consumer in one bisectable commit; representation UNCHANGED underneath and every bench pin BIT-IDENTICAL, the survey's theorem made measurable. THE MECHANISM (the survey §2.3 pick, all three sanctioned engine seams, zero new TyKind/opcode/nat/encoded vocabulary — VERSION 11 NO BUMP, nothing on the wire): (a) the alias head admits (Ident | Type) members — TypeAliasFrame parses `type Name<K, i64> = ..;` with the head members as type nodes (AliasData.params; the dump renders params only when non-empty so every plain-alias dump is byte-identical); (b) CONCRETE-ROWS-FIRST resolution — resolve_type matches a family's rows exactly-first at substitution-completing sites (parameters bind the site's argument, concrete members equal by TypeId) and expands to the row TARGET under the head's substitution BEFORE any table work (RFC 0043 §4's no-IR law: the spelled name resolves to the target's TypeId — the class instantiates only when no row matches, the C2 duplicate-type-name check lifts exactly for the concrete-shadows-generic shape, several rows may share one family name); (c) impl-target alias expansion — `impl JsonDeserialize for HashMap<K, i64>` re-targets through the row at the NODE level (the impl's generic idents bind the head's parameter members, concrete members must spell their types) and registers on the row TARGET's class template — the SAME (trait, type) pair and substitution the direct spelling produces — and plain aliases expand in impl-target position too (probe D1's lift: `impl Paint for B2` compiles, inherent + trait + orphan-origin via the expanded decl). The static-call mint rides the same law (call.rs): explicit `HashMap<str, i64>.new()` and the expected-type inference arm both re-target to the target instantiation. THE SURFACE: nmapset's row table lands (`pub type HashMap<K, i64> = PrimMapI64<K>;` + u64/f64) and the three prim classes DROP pub — nmapset-internal; the LSP std surface re-scopes to the family for real (TyDef gains is_pub, nmapset's index rows retain pub-only — the internal names leave completions/index, pinned by new negative assertions in std_surface + e2e). THE MIGRATION (every census site): json's impl matrix re-targets through the rows (`use nmapset::{HashMap, HashSet};`, the three decode rows spell HashMap<K, i64/u64/f64> and gain the general map's key_spells branch so the bytes-key diagnostic text does not move with the re-seat — dec_hm_bytes_key's pin holds verbatim); the jsonpkg fixture's three decodeJson rows re-spell same-checksum; examples/05's todo_list tables + app.rut fields/mints/rebinds re-spell, the four sidecar-i64 sites (t1 lids, atom gens, derived m, valcolumn wrapper) RE-SEAT to the val column silently-by-law as disclosed, todolist's law gate retunes to `HashMap<str, i64>`, demo's maps case + blurbs re-spell, the bench row re-spells SOURCE-ONLY (`HashMap<i32, i64> = HashMap.new()`), json.rut/web_dom/Cargo.toml/rut.toml comments reword, benches/README gains the appended phase section (history untouched); driver suites retuned per §2.5 (nmap_primmap keeps the checksum pins and drops the vacuous self-comparisons, headers disclose; nmap_valcolumn's law now pins the row spelling against the raw crossings — the re-seat's own proof). ONE CENSUS GAP FOUND AND RETUNED HONESTLY: nmap_viewkeys.rs (strings-round1's range-method suite) spells `HashMap<str, i64>` + put_range/get_range — the range surface lives ONLY on the generic class, so the rows would strand it; the suite now spells the no-row val `HashMap<str, i32>` (the nmapset-str shape it parity-matches), every pinned literal unchanged (2572351 included), header discloses; range-methods-on-the-column recorded as the new-surface menu item. THE THEOREM, MEASURED on the rebuilt release binaries: nmap-primmap checksum 734932704 + fuel 17,950,301 + heap 324 B BIT-IDENTICAL through the re-spelled source; all sibling pins exact (nmapset-int 734932704/20,703,284/1,966,551; str 1264308351/9,551,761/983,620; knucleotide 2198604/38,814,389/4,195,084; hashset 21500055/13,267,176/551); full suite 30 rows × {rut, qjs, node} exit 0 with every checksum equal to expected.json — expected.json UNTOUCHED, the pin is the gate. TWO SURVEY CLAIMS CORRECTED ON RECEIPT, not worked around: (1) the pub drop does NOT hard-fail a straggler `use nmapset::{PrimMapI64}` — the splice path is flat and the collector is vis-blind (probed, receipt under p1/) — the gate's enforcer is the Tier-1 grep plus the LSP pub-only surface filter, and a linked-surface pub gate stays future work; (2) the Tier-1 grep's survivors are exactly the two negative assertions that pin the absence (std_surface.rs, e2e-wasm.js — the gate's own enforcement code) and benches/README.md's historical perf-log sections (:2016+ — past batches' records, the strbuild not-rewritten precedent; the survey's Tier-2 list under-counted this third carve-out); everything else live greps ZERO. LSP: std_surface five-name pin retunes to the family pair + absent internals, wasm REBUILT through build:wasm (bin/rut-lsp.wasm md5 d9ca8870aa5c439e07bdb23c49dfc906), grammar corpus 86 files 0 violations, e2e through the REBUILT artifact 86 corpus files ZERO false diagnostics + 22 smokes green, vsix re-issued 0.2.3 -> 0.2.4 (rut-vscode-0.2.4.vsix md5 af9b86ba7bbd9f769234664d8551b888; both gitignored build outputs — the md5s are the receipt). GATES on this exact tree: cargo test --workspace 779 passed / 0 failed (the base count exactly — retunes in place, no test lost); wasm32-unknown-unknown check exit 0; the full bench suite green as above; demo wasm rebuilt + smoke 281 passed / 0 failed (ex-maps through the new engine, zero false diagnostics); todolist-web 93 passed / 0 failed; tree clean after the commit. Staged by EXPLICIT PATH ONLY (the parallel deploy lane's demo/package.json mod + untracked scripts/ and .wrangler/ untouched, not staged). Scratch probes + gate receipts under /tmp/opencode/batch-hashmap-surface/p1/ (probe-rows/probe-impls/probe-admit/probe-pubdrop + logs, gates-cargo-test.log, gates-wasm32.log, gates-bench-full.json, gates-benchpins.txt, gates-tier1.txt, vsix-md5.txt). Phase 2 (the close-out: the batch report + the RFC 0043 amendment verification) remains.

+++ crates/rut-driver/tests/nmap_viewkeys.rs
@@ -3,6 +3,15 @@
 //! `has_range` / `remove_range`, the key crossing as a borrowed
 //! `(parent, off, len)` BYTE range.
 //!
+//! SPELLING (the hashmap-surface batch): the range-keyed methods live
+//! on the GENERIC class, and the family rows re-seat `HashMap<str,
+//! i64>` to the val-column class — which has no range surface. This
+//! suite spells the no-row val `HashMap<str, i32>` (the nmapset-str
+//! shape this suite parity-matches anyway) so the range machinery
+//! stays testable; every pinned literal below is unchanged (the folded
+//! integers are the same). The range-surface-on-the-column question is
+//! a new-surface menu item, not a rename.
+//!
 //! Covered: the PARITY law through the wrapper — the nmapset-str churn
 //! shape spelled both ways over one parent (range methods vs
 //! slice-then-put) answers the SAME pinned checksum; lane
@@ -87,14 +96,14 @@ fn build_parent(n: i32) -> str {
 
 // the nmapset-str six-phase churn, RANGE-spelled
 fn churn_range(p: str, n: i32) -> i64 {
-    let mut m: HashMap<str, i64> = HashMap.new();
+    let mut m: HashMap<str, i32> = HashMap.new();
     let mut added: i64 = 0;
     for (let i = 0; i < n; i += 1) {
-        if (m.put_range(p, i * 6, 6, i as i64)) { added += 1; }
+        if (m.put_range(p, i * 6, 6, i)) { added += 1; }
     }
     let mut replaced: i64 = 0;
     for (let i = 0; i < n; i += 1) {
-        if (m.put_range(p, i * 6, 6, i as i64 + 3) == false) { replaced += 1; }
+        if (m.put_range(p, i * 6, 6, i + 3) == false) { replaced += 1; }
     }
     let mut sum: i64 = 0;
     let mut hits: i64 = 0;
@@ -102,7 +111,7 @@ fn churn_range(p: str, n: i32) -> i64 {
         let q = m.get_range(p, i * 6, 6);
         if (q != nil) {
             hits += 1;
-            sum = sum.wrapping_add(q);
+            sum = sum.wrapping_add(q as i64);
         }
     }
     let mut misses: i64 = 0;
@@ -119,7 +128,7 @@ fn churn_range(p: str, n: i32) -> i64 {
     }
     let mut added2: i64 = 0;
     for (let i = 0; i < n; i += 3) {
-        if (m.put_range(p, i * 6, 6, 0i64 - i as i64)) { added2 += 1; }
+        if (m.put_range(p, i * 6, 6, 0 - i)) { added2 += 1; }
     }
     let mut c: i64 = sum;
     c = c.wrapping_add(added.wrapping_mul(31));
@@ -135,14 +144,14 @@ fn churn_range(p: str, n: i32) -> i64 {
 
 // the identical churn, SLICE-spelled (the pre-phase-2 spelling)
 fn churn_slice(p: str, n: i32) -> i64 {
-    let mut m: HashMap<str, i64> = HashMap.new();
+    let mut m: HashMap<str, i32> = HashMap.new();
     let mut added: i64 = 0;
     for (let i = 0; i < n; i += 1) {
-        if (m.put(p.slice(i * 6, i * 6 + 6), i as i64)) { added += 1; }
+        if (m.put(p.slice(i * 6, i * 6 + 6), i)) { added += 1; }
     }
     let mut replaced: i64 = 0;
     for (let i = 0; i < n; i += 1) {
-        if (m.put(p.slice(i * 6, i * 6 + 6), i as i64 + 3) == false) { replaced += 1; }
+        if (m.put(p.slice(i * 6, i * 6 + 6), i + 3) == false) { replaced += 1; }
     }
     let mut sum: i64 = 0;
     let mut hits: i64 = 0;
@@ -150,7 +159,7 @@ fn churn_slice(p: str, n: i32) -> i64 {
         let q = m.get(p.slice(i * 6, i * 6 + 6));
         if (q != nil) {
             hits += 1;
-            sum = sum.wrapping_add(q);
+            sum = sum.wrapping_add(q as i64);
         }
     }
     let mut misses: i64 = 0;
@@ -167,7 +176,7 @@ fn churn_slice(p: str, n: i32) -> i64 {
     }
     let mut added2: i64 = 0;
     for (let i = 0; i < n; i += 3) {
-        if (m.put(p.slice(i * 6, i * 6 + 6), 0i64 - i as i64)) { added2 += 1; }
+        if (m.put(p.slice(i * 6, i * 6 + 6), 0 - i)) { added2 += 1; }
     }
     let mut c: i64 = sum;
     c = c.wrapping_add(added.wrapping_mul(31));
@@ -212,20 +221,20 @@ use nmapset::{ HashMap };
 
 pub fn main() -> i64 {
     let seq = "ACGTTCAGGCATXZ";
-    let mut m: HashMap<str, i64> = HashMap.new();
+    let mut m: HashMap<str, i32> = HashMap.new();
     let mut f: i64 = 0;
     // range inserts, slice reads
     if (!m.put_range(seq, 0, 3, 1)) { f += 1; }      // "ACG"
     if (!m.put_range(seq, 3, 3, 2)) { f += 2; }      // "TTC"
-    if (m.get(seq.slice(0, 3)) != 1i64) { f += 4; }
+    if (m.get(seq.slice(0, 3)) != 1) { f += 4; }
     if (!m.has(seq.slice(3, 6))) { f += 8; }
     // slice put over a range key: replace, not a second entry
-    if (m.put(seq.slice(0, 3), 10i64)) { f += 16; }
+    if (m.put(seq.slice(0, 3), 10)) { f += 16; }
     if (m.len() != 2) { f += 32; }
-    if (m.get_range(seq, 0, 3) != 10i64) { f += 64; }
+    if (m.get_range(seq, 0, 3) != 10) { f += 64; }
     // a window key off the slot grid: range stores, slice finds
     if (!m.put_range(seq, 5, 4, 3)) { f += 128; }    // "CAGG"
-    if (m.get(seq.slice(5, 9)) != 3i64) { f += 256; }
+    if (m.get(seq.slice(5, 9)) != 3) { f += 256; }
     if (!m.has_range(seq, 5, 4)) { f += 512; }
     // cross-lane removes, both directions
     if (!m.remove(seq.slice(0, 3))) { f += 1024; }
@@ -238,8 +247,8 @@ pub fn main() -> i64 {
     // the surviving range key: this is a REPLACE (answers false)
     let win = seq.slice(0, 9);
     if (m.put_range(win, 3, 3, 99) != false) { f += 32768; }
-    if (m.get_range(seq, 3, 3) != 99i64) { f += 65536; }
-    if (m.get(seq.slice(3, 6)) != 99i64) { f += 131072; }
+    if (m.get_range(seq, 3, 3) != 99) { f += 65536; }
+    if (m.get(seq.slice(3, 6)) != 99) { f += 131072; }
     if (m.len() != 1) { f += 262144; }
     return f;
 }
@@ -268,27 +277,27 @@ pub fn main() -> i64 {
         let base = 1000 + i;
         p = f"{p}{base}";
     }
-    let mut m: HashMap<str, i64> = HashMap.with_capacity(4);
+    let mut m: HashMap<str, i32> = HashMap.with_capacity(4);
     let mut fails: i64 = 0;
     let n = 500;
     for (let i = 0; i < n; i += 1) {
-        if (!m.put_range(p, i * 4, 4, i as i64 * 7 + 1)) { fails += 1; }
+        if (!m.put_range(p, i * 4, 4, i * 7 + 1)) { fails += 1; }
     }
     if (m.len() != n) { fails += 64; }
     for (let i = 0; i < n; i += 3) {
         if (!m.remove_range(p, i * 4, 4)) { fails += 2; }
     }
     for (let i = 0; i < n; i += 3) {
-        if (!m.put_range(p, i * 4, 4, i as i64 * 7 + 1)) { fails += 4; }
+        if (!m.put_range(p, i * 4, 4, i * 7 + 1)) { fails += 4; }
     }
     for (let i = 0; i < n; i += 1) {
         let q = m.get_range(p, i * 4, 4);
         if (q == nil) { fails += 8; }
-        if (q != i as i64 * 7 + 1) { fails += 16; }
+        if (q != i * 7 + 1) { fails += 16; }
     }
     // the plain lanes see the same entries
     for (let i = 0; i < n; i += 97) {
-        if (m.get(p.slice(i * 4, i * 4 + 4)) != i as i64 * 7 + 1) { fails += 32; }
+        if (m.get(p.slice(i * 4, i * 4 + 4)) != i * 7 + 1) { fails += 32; }
     }
     return fails;
 }
@@ -308,14 +317,14 @@ use nmapset::{ HashMap };
 
 pub fn main() -> i64 {
     let s = "abc";
-    let mut a: HashMap<str, i64> = HashMap.new();
-    let mut b: HashMap<str, i64> = HashMap.new();
+    let mut a: HashMap<str, i32> = HashMap.new();
+    let mut b: HashMap<str, i32> = HashMap.new();
     let mut f: i64 = 0;
     if (!a.put_range(s, 3, 0, 7)) { f += 1; }
-    if (a.get("") != 7i64) { f += 2; }
+    if (a.get("") != 7) { f += 2; }
     if (!b.put("", 9)) { f += 4; }
-    if (b.get_range(s, 0, 0) != 9i64) { f += 8; }
-    if (b.get_range(s, 3, 0) != 9i64) { f += 16; }  // any boundary window of len 0
+    if (b.get_range(s, 0, 0) != 9) { f += 8; }
+    if (b.get_range(s, 3, 0) != 9) { f += 16; }  // any boundary window of len 0
     if (a.len() != 1 || b.len() != 1) { f += 32; }
     return f;
 }
@@ -346,7 +355,7 @@ const TRAP_SRC: &str = r#"
 use nmapset::{ HashMap };
 
 pub fn main() -> i64 {
-    let mut m: HashMap<str, i64> = HashMap.new();
+    let mut m: HashMap<str, i32> = HashMap.new();
     m.put("ok", 1);
     let utf8 = "héllo";
     m.get_range(utf8, 0, 2);   // byte 2 is mid-codepoint: trap
@@ -371,7 +380,7 @@ const PAST_END_SRC: &str = r#"
 use nmapset::{ HashMap };
 
 pub fn main() -> i64 {
-    let mut m: HashMap<str, i64> = HashMap.new();
+    let mut m: HashMap<str, i32> = HashMap.new();
     m.put("ok", 1);
     let s = "abc";
     m.remove_range(s, 1, 3);   // 1 + 3 = 4 > 3 bytes: trap
